scripts/requirements.txt
flask
flask[async]
pyyaml>=5.4

psycopg2-binary==2.9.9
sqlalchemy<2.0
rasa==3.6.21

numpy
ffmpeg-python==0.2.0
openai-whisper
--extra-index-url https://download.pytorch.org/whl/cu113
torch

#claranet4

sounddevice
soundfile

#pandas
#plotly
#pygraphviz
#matplotlib
#imageio

#pyaudio
#pydub
#SpeechRecognition
#pvporcupine

#claranet4

chime
